Large commit. Was coding a better Sorter class due to spike or spikerelated task and therefore used a local repository rather than pvg's. Made the configParser singleton for accesibility and made changes to reflect this (changed test files, IOHandler) so that it access the class in a more static context.
The code has a new runner class and a new runner sorter as well as separated printer classes for readability and sanity. The comparators uses a nifty design pattern that makes them more java bean-like (for clarity and sanity in equal parts). All Tests works on my end so I feel safe in pushing this to the repository